centrenet_keras/requirements.txt
keras==2.3.1
torch==1.5.0
typing_extensions==3.7.4.2
keras-resnet==0.2.0
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==6.2.0
tensorflow-gpu==1.14.0
progressbar2